fix(tests): split TestBundleFactsLibraryManifest into its own file
tests/test_cli_compare_bundle_facts.py hit 1203 lines after the last
skip-guard fix, tripping the architecture gate's 1200-line test-file
cap (new-test-size). Split TestBundleFactsLibraryManifest out into a
sibling tests/test_cli_compare_bundle_facts_manifest.py, mirroring the
production split between compare_bundle_facts.py (CLI dispatch,
TestCompareOldBundleFacts's own file) and
workflows/bundle_facts_library_overrides.py (the manifest parser this
moved class exercises) -- the same convention test_run_plan.py's own
sibling splits already follow (tests/CLAUDE.md).

Both files verified independently: 29/29 tests pass, ruff/mypy/
architecture checks clean.
